+++ csFastFloat/FastFloatParser.cs
@@ -53,7 +53,7 @@ namespace csFastFloat
     /// <returns></returns>
     internal static float FastPath(ParsedNumberString pns)
     {
-      float value = (float)pns.mantissa;
+      float value = pns.mantissa;
       if (pns.exponent < 0)
       {
         value /= Exact_power_of_ten(-pns.exponent);
@@ -536,7 +536,7 @@ namespace csFastFloat
       }
       // If we called compute_float<binary_format<T>>(pns.exponent, pns.mantissa) and we have an invalid power (am.power2 < 0),
       // then we need to go the long way around again. This is very uncommon.
-      if (am.power2 < 0) { am = ParseLongMantissa(first, last, (byte)decimal_separator); }
+      if (am.power2 < 0) { am = ParseLongMantissa(first, last, decimal_separator); }
       result = ToFloat(pns.negative, am);
       return true;
     }
@@ -600,7 +600,7 @@ namespace csFastFloat
 
       answer.mantissa = product.high >> (upperbit + 64 - FloatBinaryConstants.mantissa_explicit_bits - 3);
 
-      answer.power2 = (int)(Utils.power((int)(q)) + upperbit - lz - FloatBinaryConstants.minimum_exponent);
+      answer.power2 = Utils.power((int)(q)) + upperbit - lz - FloatBinaryConstants.minimum_exponent;
       if (answer.power2 <= 0)
       { // we have a subnormal?
         // Here have that answer.power2 <= 0 so -answer.power2 >= 0
@@ -713,7 +713,7 @@ namespace csFastFloat
           answer.mantissa = 0;
           return answer;
         }
-        exp2 += (int)(shift);
+        exp2 += shift;
       }
       // We shift left toward [1/2 ... 1].
       while (d.decimal_point <= 0)
@@ -744,7 +744,7 @@ namespace csFastFloat
           answer.mantissa = 0;
           return answer;
         }
-        exp2 -= (int)(shift);
+        exp2 -= shift;
       }
       // We are now in the range [1/2 ... 1] but the binary format uses [1 ... 2].
       exp2--;
@@ -753,13 +753,13 @@ namespace csFastFloat
 
       while ((min_exp + 1) > exp2)
       {
-        int n = (int)((min_exp + 1) - exp2);
+        int n = (min_exp + 1) - exp2;
         if (n > max_shift)
         {
           n = max_shift;
         }
         d.decimal_right_shift(n);
-        exp2 += (int)(n);
+        exp2 += n;
       }
       if ((exp2 - min_exp) >= FloatBinaryConstants.infinite_power)
       {
@@ -769,7 +769,7 @@ namespace csFastFloat
       }
 
       int mantissa_size_in_bits = FloatBinaryConstants.mantissa_explicit_bits + 1;
-      d.decimal_left_shift((int)mantissa_size_in_bits);
+      d.decimal_left_shift(mantissa_size_in_bits);
 
       ulong mantissa = d.round();
       // It is possible that we have an overflow, in which case we need
